// src/graphql/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** An ISO-8601 encoded UTC date string */
  DateTime: { input: any; output: any; }
  /** An ISO 8601-encoded date */
  ISO8601Date: { input: any; output: any; }
  /**
   * Represents non-fractional signed whole numeric values.
   *
   * `JSInt` can represent values between -(2^53) + 1 and 2^53 - 1.
   *
   */
  JSInt: { input: any; output: any; }
  /** A blob of JSON represented as a pretty formatted string */
  JSON: { input: any; output: any; }
  /** A Pipeline identifier using a slug, and optionally negated with a leading `!` */
  PipelineSelector: { input: any; output: any; }
  /** A Team identifier using a slug, and optionally negated with a leading `!` */
  TeamSelector: { input: any; output: any; }
  /** A User identifier using a UUID, and optionally negated with a leading `!` */
  UserSelector: { input: any; output: any; }
  /** A blob of XML represented as a pretty formatted string */
  XML: { input: any; output: any; }
  /** A blob of YAML */
  YAML: { input: any; output: any; }
};
// ...
export type Organization = Node & {
  __typename?: 'Organization';
  /** Returns agent access tokens for an Organization. By default returns all tokens, whether revoked or non-revoked. */
  agentTokens?: Maybe<AgentTokenConnection>;
  agents?: Maybe<AgentConnection>;
  /** A space-separated allowlist of IP addresses that can access the organization via the GraphQL or REST API */
  allowedApiIpAddresses?: Maybe<Scalars['String']['output']>;
  /** Returns user API access tokens that can access this organization */
  apiAccessTokens: OrganizationApiAccessTokenConnection;
  auditEvents?: Maybe<OrganizationAuditEventConnection>;
  /** Returns active banners for this organization. */
  banners: OrganizationBannerConnection;
  /** Return cluster in the Organization by UUID */
  cluster?: Maybe<Cluster>;
  /** Returns clusters for an Organization */
  clusters?: Maybe<ClusterConnection>;
  /** The URL to an icon representing this organization */
  iconUrl?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  invitations?: Maybe<OrganizationInvitationConnection>;
  /** Whether teams is enabled for this organization */
  isTeamsEnabled: Scalars['Boolean']['output'];
  jobs?: Maybe<JobConnection>;
  /** Returns users within the organization */
  members?: Maybe<OrganizationMemberConnection>;
  /** Whether this organization requires 2FA to access (Please note that this is a beta feature and is not yet available to all organizations.) */
  membersRequireTwoFactorAuthentication: Scalars['Boolean']['output'];
  /** The name of the organization */
  name: Scalars['String']['output'];
  permissions: OrganizationPermissions;
  /** Return all the pipeline templates the current user has access to for this organization */
  pipelineTemplates?: Maybe<PipelineTemplateConnection>;
  /** Return all the pipelines the current user has access to for this organization */
  pipelines?: Maybe<PipelineConnection>;
  /** Whether this organization is visible to everyone, including people outside it */
  public: Scalars['Boolean']['output'];
  /** Return all the registries the current user has access to for this organization */
  registries?: Maybe<RegistryConnection>;
  /** Returns the repository providers for this organization */
  repositoryProviders?: Maybe<Array<Maybe<OrganizationRepositoryProvider>>>;
  /** API tokens with access to this organization will be automatically revoked after this many seconds of inactivity. A `null` value indicates never revoke inactive tokens. */
  revokeInactiveTokensAfter?: Maybe<RevokeInactiveTokenPeriod>;
  /** Returns rules for an Organization */
  rules?: Maybe<RuleConnection>;
  /** The slug used to represent the organization in URLs */
  slug: Scalars['String']['output'];
  /** The single sign-on configuration of this organization */
  sso?: Maybe<OrganizationSso>;
  /** Single sign on providers created for an organization */
  ssoProviders?: Maybe<SsoProviderConnection>;
  /** Return all the suite the current user has access to for this organization */
  suites?: Maybe<SuiteConnection>;
  /** Returns teams within the organization that the viewer can see */
  teams?: Maybe<TeamConnection>;
  /** Returns the resource usage data for this organization. */
  usage: UsageUnionConnection;
  /** The public UUID for this organization */
  uuid: Scalars['String']['output'];
};
// ...
export type OrganizationConnection = Connection & {
  __typename?: 'OrganizationConnection';
  count: Scalars['Int']['output'];
  edges?: Maybe<Array<Maybe<OrganizationEdge>>>;
  pageInfo?: Maybe<PageInfo>;
};

export type OrganizationEdge = {
  __typename?: 'OrganizationEdge';
  cursor: Scalars['String']['output'];
  node?: Maybe<Organization>;
};
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars['String']['output']>;
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars['Boolean']['output'];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars['Boolean']['output'];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars['String']['output']>;
};
// ...
export type PipelineConnection = Connection & {
  __typename?: 'PipelineConnection';
  count: Scalars['Int']['output'];
  edges?: Maybe<Array<Maybe<PipelineEdge>>>;
  pageInfo?: Maybe<PageInfo>;
};
// ...
export type PipelineEdge = {
  __typename?: 'PipelineEdge';
  cursor: Scalars['String']['output'];
  node?: Maybe<Pipeline>;
};
// ...
export type Query = {
  __typename?: 'Query';
  /** Find an agent by its slug */
  agent?: Maybe<Agent>;
  /** Find an agent token by its slug */
  agentToken?: Maybe<AgentToken>;
  /** Find a API Access Token code */
  apiAccessTokenCode?: Maybe<ApiAccessTokenCode>;
  /** Find an artifact by its UUID */
  artifact?: Maybe<Artifact>;
  /** Find an audit event via its uuid */
  auditEvent?: Maybe<AuditEvent>;
  /** Find a build */
  build?: Maybe<Build>;
  /** Find a GraphQL snippet */
  graphQLSnippet?: Maybe<GraphQlSnippet>;
  /** Find a build job */
  job?: Maybe<Job>;
  /** Fetches an object given its ID. */
  node?: Maybe<Node>;
  /** Find a notification service via its UUID */
  notificationService?: Maybe<NotificationService>;
  /** Find an organization */
  organization?: Maybe<Organization>;
  /** Find an organization invitation via its slug */
  organizationInvitation?: Maybe<OrganizationInvitation>;
  /** Find an organization membership via its slug */
  organizationMember?: Maybe<OrganizationMember>;
  /** Find a pipeline */
  pipeline?: Maybe<Pipeline>;
  /** Find a pipeline schedule by its slug */
  pipelineSchedule?: Maybe<PipelineSchedule>;
  /** Find a pipeline template */
  pipelineTemplate?: Maybe<PipelineTemplate>;
  /** Find a registry */
  registry?: Maybe<Registry>;
  /** Find a rule via its UUID */
  rule?: Maybe<Rule>;
  /** Find a secret via its uuid. This does not contain the value of the secret or encrypted material. */
  secret?: Maybe<Secret>;
  /** Find an sso provider either using it's slug, or UUID */
  ssoProvider?: Maybe<SsoProvider>;
  /** Find a team */
  team?: Maybe<Team>;
  /** Context of the current user using the GraphQL API */
  viewer?: Maybe<Viewer>;
};
// ...
export type Repository = {
  __typename?: 'Repository';
  /** The repository’s provider */
  provider?: Maybe<RepositoryProvider>;
  /** The git URL for this repository */
  url: Scalars['String']['output'];
};
// ...
export type User = Node & {
  __typename?: 'User';
  avatar: Avatar;
  /** If this user account is an official bot managed by Buildkite */
  bot: Scalars['Boolean']['output'];
  /** Returns builds that this user has created. */
  builds?: Maybe<BuildConnection>;
  /** The primary email for the user */
  email: Scalars['String']['output'];
  /** Does the user have a password set */
  hasPassword: Scalars['Boolean']['output'];
  id: Scalars['ID']['output'];
  /** The name of the user */
  name: Scalars['String']['output'];
  /** The public UUID of the user */
  uuid: Scalars['String']['output'];
};


/** A user */
export type UserBuildsArgs = {
  branch?: InputMaybe<Array<Scalars['String']['input']>>;
  first?: InputMaybe<Scalars['Int']['input']>;
  last?: InputMaybe<Scalars['Int']['input']>;
  metaData?: InputMaybe<Array<Scalars['String']['input']>>;
  state?: InputMaybe<Array<BuildStates>>;
};

/** Represents the current user session */
export type Viewer = Node & {
  __typename?: 'Viewer';
  authorizations?: Maybe<AuthorizationConnection>;
  builds?: Maybe<BuildConnection>;
  /** Emails associated with the current user */
  emails?: Maybe<EmailConnection>;
  /** The ID of the current user */
  id: Scalars['ID']['output'];
  notice?: Maybe<Notice>;
  organizations?: Maybe<OrganizationConnection>;
  /** The current user's permissions */
  permissions: ViewerPermissions;
  /**
   * The user's active TOTP configuration, if any.
   *
   * This field is private, requires an escalated session, and cannot be accessed via the public GraphQL API.
   */
  totp?: Maybe<Totp>;
  /** The current user */
  user?: Maybe<User>;
};
// ...
export type GetViewerQuery = { __typename?: 'Query', viewer?: { __typename?: 'Viewer', id: string, user?: { __typename?: 'User', id: string, uuid: string, name: string, email: string } | null } | null };

export type GetOrganizationsQueryVariables = Exact<{ [key: string]: never; }>;


export type GetOrganizationsQuery = { __typename?: 'Query', viewer?: { __typename?: 'Viewer', organizations?: { __typename?: 'OrganizationConnection', edges?: Array<{ __typename?: 'OrganizationEdge', node?: { __typename?: 'Organization', id: string, name: string, slug: string } | null } | null> | null } | null } | null };

export type GetPipelinesQueryVariables = Exact<{
  organizationSlug: Scalars['ID']['input'];
  first?: InputMaybe<Scalars['Int']['input']>;
  after?: InputMaybe<Scalars['String']['input']>;
}>;


export type GetPipelinesQuery = { __typename?: 'Query', organization?: { __typename?: 'Organization', pipelines?: { __typename?: 'PipelineConnection', edges?: Array<{ __typename?: 'PipelineEdge', node?: { __typename?: 'Pipeline', uuid: string, id: string, name: string, slug: string, description?: string | null, url: string, repository?: { __typename?: 'Repository', url: string } | null } | null } | null> | null, pageInfo?: { __typename?: 'PageInfo', hasNextPage: boolean, endCursor?: string | null } | null } | null } | null };

export type GetBuildsQueryVariables = Exact<{
  pipelineSlug: Scalars['String']['input'];
  organizationSlug: Scalars['ID']['input'];
  first?: InputMaybe<Scalars['Int']['input']>;
}>;


export type GetBuildsQuery = { __typename?: 'Query', organization?: { __typename?: 'Organization', pipelines?: { __typename?: 'PipelineConnection', edges?: Array<{ __typename?: 'PipelineEdge', node?: { __typename?: 'Pipeline', builds?: { __typename?: 'BuildConnection', edges?: Array<{ __typename?: 'BuildEdge', node?: { __typename?: 'Build', id: string, number: number, url: string, state: BuildStates, message?: string | null, commit: string, branch: string, createdAt?: any | null, startedAt?: any | null, finishedAt?: any | null } | null } | null> | null, pageInfo?: { __typename?: 'PageInfo', hasNextPage: boolean, endCursor?: string | null } | null } | null } | null } | null> | null } | null } | null };

export type GetViewerBuildsQueryVariables = Exact<{
  first: Scalars['Int']['input'];
}>;


export type GetViewerBuildsQuery = { __typename?: 'Query', viewer?: { __typename?: 'Viewer', builds?: { __typename?: 'BuildConnection', edges?: Array<{ __typename?: 'BuildEdge', node?: { __typename?: 'Build', id: string, number: number, state: BuildStates, url: string, createdAt?: any | null, branch: string, message?: string | null, pipeline: { __typename?: 'Pipeline', name: string, slug: string }, organization: { __typename?: 'Organization', name: string, slug: string } } | null } | null> | null, pageInfo?: { __typename?: 'PageInfo', hasNextPage: boolean, endCursor?: string | null } | null } | null } | null };